bm_ModelingOptions output object class

Class returned by bm_ModelingOptions (a list of BIOMOD.options.dataset more exactly), and used by BIOMOD_Modeling

# S4 method for class 'character,character'
BIOMOD.options.default(mod, typ, pkg, fun)

Arguments

mod

a character corresponding to the model name to be computed, must be either ANN, CTA, FDA, GAM, GBM, GLM, MARS, MAXENT, MAXNET, RF, SRE, XGBOOST

typ

a character corresponding to the data type to be used, must be either binary, binary.PA, abundance, compositional

pkg

a character corresponding to the package containing the model function to be called

fun

a character corresponding to the model function name to be called

Slots

model

a character corresponding to the model

type

a character corresponding to the data type (binary, binary.PA, abundance, compositional)

package

a character corresponding to the package containing the model function to be called

func

a character corresponding to the model function name to be called

args.names

a vector containing character corresponding to the model function arguments

args.default

a list containing for each dataset the default values for all arguments listed in args.names

See also

Author

Maya Guéguen

Examples

showClass("BIOMOD.options.default")