BIOMOD_EnsembleModeling() output object class

Classes used by BIOMOD_Modeling and BIOMOD_EnsembleModeling to build their output object (see BIOMOD.models.out objects)

Details

BIOMOD.stored.data is the basic object containing the slots inMemory and link.
All listed classes below are derived from BIOMOD.stored.data, and contain a val slot of specific type :

BIOMOD.stored.array : val is an array

BIOMOD.stored.data.frame : val is a data.frame

BIOMOD.stored.SpatRaster : val is a RasterStack

BIOMOD.stored.files : val is a character

BIOMOD.stored.formated.data : val is a BIOMOD.formated.data object

BIOMOD.stored.models.options : val is a BIOMOD.models.options object

BIOMOD.stored.models.out : val is a BIOMOD.models.out object

Slots

inMemory

a logical defining whether the val slot has been loaded in memory or not

link

a character containing the file name of the saved val slot

val

an object of type depending on the BIOMOD.stored.[...] class (see Details)

See also

Author

Damien Georges

Examples

showClass("BIOMOD.stored.data")
showClass("BIOMOD.stored.array") 
showClass("BIOMOD.stored.data.frame") 
showClass("BIOMOD.stored.SpatRaster") 
showClass("BIOMOD.stored.files") 
showClass("BIOMOD.stored.formated.data") 
showClass("BIOMOD.stored.models.options") 
showClass("BIOMOD.stored.models.out")