Format input data, and select pseudo-absences if wanted, for usage in biomod2

This function gathers together all input data needed (xy, presences/absences, explanatory variables, and the same for evaluation data if available) to run biomod2 models. It allows to select pseudo-absences if no absence data is available, with different strategies (see Details).

BIOMOD_FormatingData(
  resp.var,
  expl.var,
  resp.xy = NULL,
  resp.name = NULL,
  eval.resp.var = NULL,
  eval.expl.var = NULL,
  eval.resp.xy = NULL,
  PA.nb.rep = 0,
  PA.nb.absences = 1000,
  PA.strategy = "random",
  PA.dist.min = 0,
  PA.dist.max = NULL,
  PA.sre.quant = 0.025,
  PA.table = NULL,
  na.rm = TRUE
)

Arguments

resp.var
a vector, SpatialPoints (if presence-only) or SpatialPointsDataFrame object containing binary data (0 : absence, 1 : presence, NA : indeterminate) for a single species that will be used to build the species distribution model(s)

expl.var
a matrix, data.frame, SpatialPointsDataFrame or RasterStack object containing the explanatory variables (in columns or layers) that will be used to build the species distribution model(s)

resp.xy
(optional, default NULL)
If resp.var is a vector, a 2-columns matrix containing the corresponding X and Y coordinates that will be used to build the species distribution model(s)

resp.name
(optional, default NULL)
A character containing the species name.

eval.resp.var
(optional, default NULL)
A vector, SpatialPoints (if presence-only) or SpatialPointsDataFrame object containing binary data (0 : absence, 1 : presence, NA : indeterminate) for a single species that will be used to evaluate the species distribution model(s) with independent data

eval.expl.var
(optional, default NULL)
A matrix, data.frame, SpatialPointsDataFrame or RasterStack object containing the explanatory variables (in columns or layers) that will be used to evaluate the species distribution model(s) with independent data

eval.resp.xy
(optional, default NULL)
If resp.var is a vector, a 2-columns matrix containing the corresponding X and Y coordinates that will be used to evaluate the species distribution model(s) with independent data

PA.nb.rep
(optional, default 0)
If pseudo-absence selection, the number of sets (repetitions) of pseudo-absence points that will be drawn

PA.nb.absences
(optional, default 0)
If pseudo-absence selection, the number of pseudo-absence points that will be selected for each pseudo-absence repetition (true absences included)

PA.strategy
(optional, default NULL)
If pseudo-absence selection, a character defining the strategy that will be used to select the pseudo-absence points (must be random, sre, disk or user.defined, see Details)

PA.dist.min
(optional, default 0)
If pseudo-absence selection and PA.strategy = 'sre', a numeric defining the minimal distance to presence points used to make the disk pseudo-absence selection (in meters, see Details)

PA.dist.max
(optional, default 0)
If pseudo-absence selection and PA.strategy = 'sre', a numeric defining the maximal distance to presence points used to make the disk pseudo-absence selection (in meters, see Details)

PA.sre.quant
(optional, default 0)
If pseudo-absence selection and PA.strategy = 'sre', a numeric between 0 and 1 defining the quantile used to make the sre pseudo-absence selection (see Details)

PA.table
(optional, default NULL)
If pseudo-absence selection and PA.strategy = 'user.defined', a matrix or data.frame with as many rows as resp.var values, as many columns as PA.nb.rep, and containing TRUE or FALSE values defining which points will be used to build the species distribution model(s) for each repetition (see Details)

na.rm
(optional, default TRUE)
A logical value defining whether points having one or several missing values for explanatory variables should be removed from the analysis or not

Value

A BIOMOD.formated.data object that can be used to build species distribution model(s) with the BIOMOD_Modeling function.
print and plot functions are available to have a summary of the created object.

Details

This function gathers and formats all input data needed to run biomod2 models. It supports different kind of inputs (e.g. matrix, SpatialPoints, RasterStack) and provides different methods to select pseudo-absences if needed.

Concerning explanatory variables and XY coordinates :

if rasterLayer or rasterStack provided (for expl.var or eval.expl.var), biomod2 will extract the corresponding values from XY coordinates provided (resp.xy or eval.resp.xy respectively).
Be sure to give the XY coordinates in the same projection system than the raster objects !

if SpatialPointsDataFrame provided (for resp.var or eval.resp.var), the same rule applies (same projection system !).

if data.frame provided (for expl.var or eval.expl.var), biomod2 will simply merge it (cbind) with resp.var without considering XY coordinates.
Be sure to give explanatory and response values in the same row order !

Concerning pseudo-absence selection :

if both presence and absence data are available, and there is enough absences : set PA.nb.rep = 0 and no pseudo-absence will be selected.

if no absence data (or not enough) is available, several pseudo-absence repetitions are recommended (to estimate the effect of pseudo-absence selection), as well as high number of pseudo-absence points.
Be sure not to select more pseudo-absence points than maximum number of pixels in the studied area !

Response variable

biomod2 models single species at a time (no multi-species). Hence, resp.var must be a uni-dimensional object (either a vector, a one-column matrix/data.frame/SpatialPointsDataFrame, or a SpatialPoints object if presence-only), containing values among :

1 : presences

0 : true absences (if any)

NA : no information point (might be used to select pseudo-absences if any)

If no true absences are available, pseudo-absence selection must be done.
If resp.var is a non-spatial object (vector, matrix or data.frame), XY coordinates must be provided through resp.xy.
If pseudo-absence points are to be selected, NA points must be provided in order to select pseudo-absences among them.

Explanatory variables

Factorial variables are allowed, but might lead to some pseudo-absence strategy or models omissions (e.g. sre).

Evaluation data

Although biomod2 provides tools to automatically divide dataset into calibration and validation parts through the modeling process (see NbRunEval and DataSplit parameters in BIOMOD_Modeling function ; or BIOMOD_CrossValidation function), it is also possible (and strongly advised) to directly provide two independent datasets, one for calibration and one for validation.

Pseudo-absence selection

If no true absences are available, pseudo-absences must be selected from the background data, meaning data there is no information whether the species of interest occurs or not. It corresponds either to the remaining pixels of the expl.var (if provided as a rasterStack) or to the points identified as NA in resp.var (if expl.var provided as a matrix or data.frame).
Several methods are available to do this selection :

random

all points of initial background are pseudo-absence candidates. PA.nb.absences are drawn randomly, for each PA.nb.rep requested.

sre

pseudo-absences have to be selected in conditions (combination of explanatory variables) that differ in a defined proportion (PA.sre.quant) from those of presence points. A Surface Range Envelop model is first run over the species of interest, and pseudo-absences are selected outside this envelop.
This case is appropriate when all the species climatic niche has been sampled, otherwise it may lead to over-optimistic model evaluations and predictions !

disk

pseudo-absences are selected within circles around presence points defined by PA.dist.min and PA.dist.max distance values (in meters). It allows to select pseudo-absence points that are not too close to (avoid same niche and pseudo-replication) or too far (localized sampling strategy) from presences.

user.defined

pseudo-absences are defined in advance and given as data.frame through the PA.table parameter.

See also

Author

Damien Georges, Wilfried Thuiller

Examples

# species occurrences
myFile <- system.file("external/species/mammals_table.csv", package="biomod2")
DataSpecies <- read.csv(myFile, row.names = 1)
head(DataSpecies)

# the name of studied species
myRespName <- 'GuloGulo'

# the presence/absences data for our species
myResp <- as.numeric(DataSpecies[, myRespName])

# the XY coordinates of species data
myRespXY <- DataSpecies[, c("X_WGS84", "Y_WGS84")]

# Environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
myFiles = paste0("external/bioclim/current/bio", c(3, 4, 7, 11, 12), ".grd")
myExpl = raster::stack(system.file(myFiles[1], package = "biomod2"),
                       system.file(myFiles[2], package = "biomod2"),
                       system.file(myFiles[3], package = "biomod2"),
                       system.file(myFiles[4], package = "biomod2"),
                       system.file(myFiles[5], package = "biomod2"))

# 1. Formating Data
myBiomodData <- BIOMOD_FormatingData(resp.var = myResp,
                                     expl.var = myExpl,
                                     resp.xy = myRespXY,
                                     resp.name = myRespName)

myBiomodData
plot(myBiomodData)