Project a range of calibrated species distribution models onto new environment

This function allows to project a range of models built with the BIOMOD_Modeling function onto new environmental data (which can represent new areas, resolution or time scales for example).

BIOMOD_Projection(
  modeling.output,
  proj.name,
  new.env,
  new.env.xy = NULL,
  chosen.models = "all",
  binary.meth = NULL,
  filtered.meth = NULL,
  compress = TRUE,
  build.clamping.mask = TRUE,
  ...
)

Arguments

modeling.output
a BIOMOD.models.out object returned by the BIOMOD_Modeling function

proj.name
a character corresponding to the name (ID) of the projection set (a new folder will be created within the simulation folder with this name)

new.env
a matrix, data.frame or RasterStack object containing the new explanatory variables (in columns or layers, with names matching the variables names given to the BIOMOD_FormatingData function to build modeling.output) that will be used to project the species distribution model(s)

new.env.xy
(optional, default NULL)
If new.env is a matrix or a data.frame, a 2-columns matrix containing the corresponding X and Y coordinates that will be used to project the species distribution model(s)

chosen.models
a vector containing model names to be kept, must be either all or a sub-selection of model names

binary.meth
(optional, default NULL)
A vector containing evaluation metric names to be used to transform prediction values into binary values based on models evaluation scores obtained with the BIOMOD_Modeling function. Must be among all (same evaluation metrics than those of modeling.output) or ROC, TSS, KAPPA, ACCURACY, BIAS, POD, FAR, POFD, SR, CSI, ETS, HK, HSS, OR, ORSS

filtered.meth
(optional, default NULL)
A vector containing evaluation metric names to be used to transform prediction values into filtered values based on models evaluation scores obtained with the BIOMOD_Modeling function. Must be among all (same evaluation metrics than those of modeling.output) or ROC, TSS, KAPPA, ACCURACY, BIAS, POD, FAR, POFD, SR, CSI, ETS, HK, HSS, OR, ORSS

compress
(optional, default TRUE)
A logical or a character value defining whether and how objects should be compressed when saved on hard drive. Must be either TRUE, FALSE, xz or gzip (see Details)

build.clamping.mask
(optional, default TRUE)
A logical value defining whether a clamping mask should be built and saved on hard drive or not (see Details)

...
(optional, see Details)

Value

A BIOMOD.projection.out object containing models projections, or links to saved outputs. Models projections are stored out of R (for memory storage reasons) in proj.name folder created in the current working directory :

the output is a 4-dimensional array if new.env is a matrix or a data.frame

it is a rasterStack if new.env is a rasterStack (or several rasterLayer objects, if new.env is too large)

raw projections, as well as binary and filtered projections (if asked), are saved in the proj.name folder

Details

###############################################################################################

If chosen.models = 'all', projections are done for all evaluation and pseudo absences runs if applicable. These projections may be used later by the BIOMOD_EnsembleForecasting function.

If build.clamping.mask = TRUE, a raster file will be saved within the projection folder. This mask values will correspond to the number of variables in each pixel that are out of their calibration / training range, identifying locations where predictions are uncertain.

... can take the following values :

omit.na : a logical value defining whether all not fully referenced environmental points will get NA as predictions or not

on_0_1000 : a logical value defining whether 0 - 1 probabilities are to be converted to 0 - 1000 scale to save memory on backup

do.stack : a logical value defining whether all projections are to be saved as one RasterStack object or several RasterLayer files (the default if projections are too heavy to be all loaded at once in memory)

keep.in.memory : a logical value defining whether all projections are to be kept loaded at once in memory, or only links pointing to hard drive are to be returned

output.format : a character value corresponding to the projections saving format on hard drive, must be either .grd, .img or .RData (the default if new.env is given as matrix or data.frame)

silent : a logical value defining whether console outputs are to be printed or not

See also

Author

Wilfried Thuiller, Damien Georges

Examples

# species occurrences
DataSpecies <- read.csv(system.file("external/species/mammals_table.csv", package="biomod2"), row.names = 1)
head(DataSpecies)

# the name of studied species
myRespName <- 'GuloGulo'

# the presence/absences data for our species
myResp <- as.numeric(DataSpecies[, myRespName])

# the XY coordinates of species data
myRespXY <- DataSpecies[, c("X_WGS84", "Y_WGS84")]

# Environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
myFiles = paste0("external/bioclim/current/bio", c(3, 4, 7, 11, 12), ".grd")
myExpl = raster::stack(system.file(myFiles[1], package = "biomod2"),
                       system.file(myFiles[2], package = "biomod2"),
                       system.file(myFiles[3], package = "biomod2"),
                       system.file(myFiles[4], package = "biomod2"),
                       system.file(myFiles[5], package = "biomod2"))

# 1. Formating Data
myBiomodData <- BIOMOD_FormatingData(resp.var = myResp,
                                     expl.var = myExpl,
                                     resp.xy = myRespXY,
                                     resp.name = myRespName)

# 2. Defining Models Options using default options.
myBiomodOption <- BIOMOD_ModelingOptions()

# 3. Doing Modelisation
myBiomodModelOut <- BIOMOD_Modeling(myBiomodData,
                                    models = c('SRE','RF'),
                                    models.options = myBiomodOption,
                                    NbRunEval = 2,
                                    DataSplit = 80,
                                    VarImport = 3,
                                    models.eval.meth = c('TSS','ROC'),
                                    do.full.models = FALSE,
                                    modeling.id = 'test')

# 4.1 Projecting on current environmental conditions
myBiomodProjection <- BIOMOD_Projection(modeling.output = myBiomodModelOut,
                                        new.env = myExpl,
                                        proj.name = 'current',
                                        chosen.models = 'all',
                                        binary.meth = 'TSS',
                                        compress = FALSE,
                                        build.clamping.mask = FALSE)

if (FALSE) {
# 4.2 Projecting on future environmental conditions

# Environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
myFiles = paste0("external/bioclim/future/bio", c(3, 4, 7, 11, 12), ".grd")
myExplFuture = raster::stack(system.file(myFiles[1], package = "biomod2"),
                             system.file(myFiles[2], package = "biomod2"),
                             system.file(myFiles[3], package = "biomod2"),
                             system.file(myFiles[4], package = "biomod2"),
                             system.file(myFiles[5], package = "biomod2"))

myBiomodProjectionFuture <- BIOMOD_Projection(modeling.output = myBiomodModelOut,
                                              new.env = myExplFuture,
                                              proj.name = 'future',
                                              chosen.models = 'all',
                                              binary.meth = 'TSS',
                                              compress = FALSE,
                                              build.clamping.mask = TRUE)

# print summary and plot projections
myBiomodProjectionFuture
plot(myBiomodProjectionFuture)
}

###############################################################################################