Produce summary outputs from a simulation folder

This function allows to produce summary report or ODMAP table from a biomod2 simulation folder.

BIOMOD_Report(
  bm.out,
  strategy = "report",
  params.color = list(color1 = "#eb4034", color2 = "#e0918b", color3 = "#658f70"),
  params.ODMAP = list(O.mod.objective = NULL, O.boundary = NULL, O.obs.type = NULL,
    O.pred.type = NULL, D.eco.level = NULL, D.samp.design = NULL)
)

Arguments

bm.out

a BIOMOD.formated.data or BIOMOD.formated.data.PA object returned by the BIOMOD_FormatingData function ; or a BIOMOD.models.out or BIOMOD.ensemble.models.out object that can be obtained with the BIOMOD_Modeling or BIOMOD_EnsembleModeling functions

strategy

a character defining the type of summary file that will be produced, must be report, ODMAP or code (see Details)

params.color

a list containing 3 color values to custom the reports

params.ODMAP

a list containing values of some ODMAP fields to be filled in from pre-existing choices (see Details)

Value

A standardized .html file obtained from an Rmarkdown template, and a .csv table in the case of ODMAP report.

Details

This function gathers and formats all objects contained in one biomod2 modeling folder to produce, based on Rmarkdown templates, standardized reports to help the user :

summarize its modeling results

share them through standardized informations through ODMAP protocol

provide reproducible code

Type of report

Different data types are available, and require different values :

report

biomod2 provides functions to summarize the information, such as print, plot or summary methods adapted to BIOMOD.[...].out objects, as well as get_[...] and bm_Plot[...] functions. All these are called here and applied to objects contained in the provided modeling folder.

ODMAP

following Zurell et al. 2020, ODMAP (Overview, Data, Model, Assessment and Prediction) protocol aims to standardize documentation of modeling to help improve both transparency and reproducibility of results. ODMAP v1.0 website provides an application to fill this type of report. biomod2 tries here to help one user to pre-fill the fields of this protocol.

code

call slot contained within BIOMOD.formated.data, BIOMOD.models.out, BIOMOD.projection.out and BIOMOD.ensemble.models.out objects keep in memory the R command used to obtain them. All these calls are gathered here in one summary file.

References

Zurell D, Franklin J, König C, Bouchet PJ, Serra-Diaz JM, Dormann CF, Elith J, Fandos Guzman G, Feng X, Guillera-Arroita G, Guisan A, Leitão PJ, Lahoz-Monfort JJ, Park DS, Peterson AT, Rapacciuolo G, Schmatz DR, Schröder B, Thuiller W, Yates KL, Zimmermann NE, Merow C (2020) A standard protocol for describing species distribution models. Ecography 43: 1261-1277. DOI: 10.1111/ecog.04960

See also

Author

Maya Gueguen

Examples

library(terra)