Abundance to build test SDM

A dataset covering France with abundance data for 20 bird species. Data comes from the French Breeding Bird Survey (STOC program) of Vigie-Nature (Fontaine et al. 2020). The original dataset contains 2,948 sites in which bird species have been observed from 2001 to 2024. Original raw data is not provided here, but might be made available on request by contacting Benoît Fontaine (benoit.fontaine [at] mnhn.fr).

DataSTOC

Format

A data.frame object with 2006 rows and 30 columns (10 variables and 20 species):

site

Observation site ID

period

yearly average headcounts have been computed over 2 times periods: 2006-2011 and 2012-2017

X_WGS84

Longitude

Y_WGS84

Latitude

temp

annual mean temperature (CHELSA)

precip

annual mean precipitation (CHELSA)

cover_agri

percentage of agricultural habitat cover around each point (European Union’s Copernicus Land Monitoring Service information)

cover_water

percentage of water habitat cover around each point (European Union’s Copernicus Land Monitoring Service information)

cover_wet

percentage of wetland habitat cover around each point (European Union’s Copernicus Land Monitoring Service information)

sdiv_hab

Shannon habitat diversity (CORINE)

Alauda.arvensis

abundance data for Eurasian skylark

Cettia.cetti

abundance data for Cetti's warbler

Coccothraustes.coccothraustes

abundance data for hawfinch

Cuculus.canorus

abundance data for common cuckoo

Emberiza.calandra

abundance data for corn bunting

Emberiza.citrinella

abundance data for yellowhammer

Erithacus.rubecula

abundance data for European robin

Fulica.atra

abundance data for Eurasian coot

Luscinia.megarhynchos

abundance data for common nightingale

Passer.domesticus

abundance data for house sparrow

Perdix.perdix

abundance data for grey partridge

Periparus.ater

abundance data for coal tit

Phylloscopus.bonelli

abundance data for western Bonelli's warbler

Phylloscopus.trochilus

abundance data for willow warbler

Regulus.regulus

abundance data for goldcrest

Serinus.serinus

abundance data for European serin

Sitta.europaea

abundance data for Eurasian nuthatch

Streptopelia.decaocto

abundance data for Eurasian collared dove

Sylvia.melanocephala

abundance data for Sardinian warbler

Turdus.philomelos

abundance data for song thrush

References

STOC EPS - Vigie-Nature (2025). French Breeding Bird Monitoring Scheme. Muséum National d’Histoire Naturelle (MNHN) - Office Français pour la Biodiversité (OFB) - Ligue pour la Protection des Oiseaux (LPO).

Fontaine B, Moussy C, Chiffard Carricaburu J, Dupuis J, Corolleur E, Schmaltz L, Lorrillière R, Loïs G, Gaudard C (2020). Suivi des oiseaux communs en France 1989-2019 : 30 ans de suivis participatifs. MNHN - Centre d'Ecologie et des Sciences de la Conservation, LPO BirdLife France - Service Connaissance, Ministère de la Transition écologique et solidaire. 46 pp.

Brun P, Zimmermann NE, Hari C, Pellissier L, Karger DN (2022). CHELSA-BIOCLIM+ A novel set of global climate-related predictors at kilometre-resolution. EnviDat. doi:10.16904/envidat.332

CORINE Land Cover 2018 (raster 100 m), Europe, 6-yearly - version 2020_20u1, May 2020. European Environment Agency. doi:10.2909/960998c1-1870-4e82-8051-6485205ebbac