Single models package and functions

A data.frame containing for each single model available in biomod2 the package and functions to be called.

ModelsTable

Format

A data.frame object with 12 rows and 5 variables:

model

all single models that can be computed in biomod2

type

data type associated to the models

package

R package used

func

function used in the R package

train

function called by caret for the tuning

All single models available are the following :

ANN (nnet)

CTA (rpart)

DNN (cito)

FDA (fda)

GAM (gam, gam or bam)

GBM (gbm)

GLM (glm)

MARS (earth)

MAXENT (see Maxent website)

MAXNET (maxnet)

RF (randomForest)

RFd (randomForest downsampled)

SRE (bm_SRE)

XGBOOST (xgboost)