Cross-validation for Neural Networks

This internal biomod2 function allows the user to compute cross-validation for neural networks in ANN model (see nnet and BIOMOD_Modeling).

bm_CVnnet(
  Input,
  Target,
  size = c(2, 4, 6, 8),
  decay = c(0.001, 0.01, 0.05, 0.1),
  maxit = 200,
  nbCV = 5,
  weights = NULL,
  seedval = 555
)

Arguments

Input

complete dataset with explanatory variables

Target

calibration dataset with observed presence / absence

size

(see parameter ANN$size in BIOMOD_ModelingOptions)

decay

(see parameter ANN$decay in BIOMOD_ModelingOptions)

maxit

(see parameter ANN$maxit in BIOMOD_ModelingOptions)

nbCV

(see parameter ANN$nbCV in BIOMOD_ModelingOptions)

weights

a vector of numeric values corresponding to weights over calibration lines

seedval

an integer value corresponding to the new seed value to be set

Value

A data.frame containing the following elements :

Size : the size

Decay : the decay value

AUC : the corresponding Area Under Curve

See also

Author

Damien Georges