Calculate the best score according to a given evaluation method

This internal biomod2 function allows the user to find the threshold to convert continuous values into binary ones leading to the best score for a given evaluation metric.

bm_FindOptimStat(
  Stat = "TSS",
  Fit,
  Obs,
  Nb.thresh.test = 100,
  Fixed.thresh = NULL
)

Arguments

Stat
a character corresponding to the evaluation metric to be used, must be either ROC, TSS, KAPPA, ACCURACY, BIAS, POD, FAR, POFD, SR, CSI, ETS, HK, HSS, OR or ORSS

Fit
a vector of fitted values (continuous)

Obs
a vector of observed values (binary, 0 or 1)

Nb.thresh.test
an integer corresponding to the number of thresholds to be tested over the range of fitted values

Fixed.thresh
(optional, default NULL)
An integer corresponding to the only threshold value to be tested

Value

A 1 row x 4 columns matrix containing :

best.iter : the best score obtained for the chosen evaluation metric

cutoff : the associated cut-off used to transform the continuous values into binary

sensibility : the sensibility obtained on fitted values with this threshold

specificity : the specificity obtained on fitted values with this threshold

Details

###############################################################################################

Please refer to BIOMOD_Modeling to get more information about these evaluation metrics.

Note that if a value is given to Fixed.thresh, no optimisation will be done., and only the score for this threshold will be returned.

See also

Author

Damien Georges

Examples

## generate a binary vector
a <- sample(c(0, 1), 100, replace = TRUE)

## random drawing
b <- runif(100, min = 0, max = 1000)
bm_FindOptimStat(Stat = 'TSS', Fit = b, Obs = a)

## biased drawing
BiasedDrawing <- function(x, m1 = 300, sd1 = 200, m2 = 700, sd2 = 200) {
  return(ifelse(x < 0.5, rnorm(1, m1, sd1), rnorm(1, m2, sd2)))
}
c <- sapply(a, BiasedDrawing)
bm_FindOptimStat(Stat = 'TSS', Fit = c, Obs = a, Nb.thresh.test = 100)

###############################################################################################