Analyze the residuals of the single models

This function return differents graphs to help analyse the single models.

bm_ModelAnalysis(
  bm.mod,
  models.chosen = "all",
  do.plot = TRUE,
  color.by = "full.name",
  do.Rpredicted = FALSE
)

Arguments

bm.mod

a BIOMOD.models.out

models.chosen

a vector containing model names to be kept, must be either all or a sub-selection of model names that can be obtained with the get_built_models function

do.plot

a logical, print the plot or not

color.by

a character between "full.name", "species", "PA", "RUN" or "algo"

do.Rpredicted

a logical : do Rpredicted have to be computed or no ?

Value

a plot residuals ~ observations to detect outliers

a histogramm of the residuals

a qqplot of the residuals

plot residuals ~ fitted

a plot of the different Rsquared available (Rsquared, Rsquared_aj, Psquared)

Details

Only the models coherent with a residuals analysis will be selected for some plots.

Author

Hélène Blancheteau