Plot mean evaluation scores

This function represents mean evaluation scores of species distribution models, from BIOMOD.models.out or BIOMOD.ensemble.models.out objects that can be obtained from BIOMOD_Modeling or BIOMOD_EnsembleModeling functions. Scores are represented according to 2 different evaluation methods, and models can be grouped (see Details).

bm_PlotEvalMean(
  modeling.output,
  eval.metric = NULL,
  group.by = "algo",
  plot = TRUE,
  ...
)

Arguments

modeling.output

a BIOMOD.models.out or BIOMOD.ensemble.models.out object that can be obtained from BIOMOD_Modeling or BIOMOD_EnsembleModeling functions

eval.metric

a vector containing evaluation metric names to be used, must be among ROC, TSS, KAPPA, ACCURACY, BIAS, POD, FAR, POFD, SR, CSI, ETS, HK, HSS, OR, ORSS

group.by

a character corresponding to the way kept models will be combined to compute mean and sd evaluation scores, must be among model, algo, run, dataset

plot

(optional, default TRUE)
A logical value defining whether the plot is to be rendered or not

...

some additional arguments (see Details)

Value

A ggplot object representing mean and standard deviation of evaluation scores according to 2 different evaluation methods.

Details

... can take the following values :

xlim : an integer corresponding to the x maximum limit to represent

ylim : an integer corresponding to the y maximum limit to represent

main : a character corresponding to the graphic title

col : a vector containing new color values

See also

Author

Damien Georges, Maya Gueguen

Examples