Select pseudo-absences

This internal biomod2 function allows to select pseudo-absences according to 4 different methods : random, sre, disk or user.defined (see Details).

bm_PseudoAbsences(
  sp,
  env,
  nb.repet = 1,
  strategy = "random",
  distMin = 0,
  distMax = NULL,
  nb.points = NULL,
  quant.SRE = 0,
  PA.table = NULL
)

bm_PseudoAbsences_user.defined(sp, env, ...)

# S4 method for ANY,SpatialPointsDataFrame
bm_PseudoAbsences_user.defined(sp, env, pa.table)

# S4 method for ANY,RasterStack
bm_PseudoAbsences_user.defined(sp, env, pa.table)

bm_PseudoAbsences_random(sp, env, ...)

# S4 method for ANY,SpatialPointsDataFrame
bm_PseudoAbsences_random(sp, env, nb.points, nb.repet)

# S4 method for ANY,RasterStack
bm_PseudoAbsences_random(sp, env, nb.points, nb.repet)

bm_PseudoAbsences_sre(sp, env, ...)

# S4 method for ANY,SpatialPointsDataFrame
bm_PseudoAbsences_sre(sp, env, quant.SRE, nb.points, nb.repet)

# S4 method for ANY,RasterStack
bm_PseudoAbsences_sre(sp, env, quant.SRE, nb.points, nb.repet)

bm_PseudoAbsences_disk(sp, env, ...)

# S4 method for ANY,SpatialPointsDataFrame
bm_PseudoAbsences_disk(sp, env, distMin, distMax, nb.points, nb.repet)

# S4 method for ANY,RasterStack
bm_PseudoAbsences_disk(sp, env, distMin, distMax, nb.points, nb.repet)

Arguments

sp

the species observations

env

the explanatory variables

nb.repet

the number of repetitions

strategy

a character corresponding to the pseudo-absence selection strategy, must be among random, sre, disk or user.defined

distMin

(optional, default 0)
If strategy = 'disk', the minimum distance away from presence points to select pseudo-absences

distMax

(optional, default NULL)
If strategy = 'disk', the maximum distance away from presence points to select pseudo-absences

nb.points

(optional, default NULL)
If strategy is random, sre or disk, the number of pseudo-absences to be selected

quant.SRE

(optional, default 0)
If strategy = 'sre', a numeric between 0 and 0.5 corresponding to the most extreme value for each variable not to be taken into account for determining the tolerance boundaries of the considered species (see bm_SRE)

PA.table

(optional, default NULL)
If strategy = 'user.defined', a matrix or data.frame with as many rows as resp.var values, as many columns as PA.nb.rep, and containing TRUE or FALSE values defining which points will be used to build the species distribution model(s) for each repetition (see BIOMOD_FormatingData)

Value

A list containing the following elements :

xy : the coordinates of the species observations

sp : the values of the species observations (0, 1 or NA)

env : the explanatory variables

pa.tab : the corresponding table of selected pseudo-absences (indicated by TRUE or FALSE)

Details

Concerning random selection :

Concerning SRE selection :

Concerning disk selection :

Concerning user defined selection :

See also

Author

Wilfried Thuiller, Damien Georges