Loop to compute all single species distribution models

This internal biomod2 function allows the user to compute all single species distribution models (asked by the BIOMOD_Modeling function).

bm_RunModelsLoop(
  X,
  modeling.id,
  Model,
  Options,
  VarImport,
  mod.eval.method,
  SavePred = TRUE,
  scal.models = TRUE
)

Arguments

X

a BIOMOD.formated.data or BIOMOD.formated.data.PA object returned by the BIOMOD_FormatingData function

modeling.id

a character corresponding to the name (ID) of the simulation set (a random number by default)

Model

a character corresponding to the model name to be computed, must be either GLM, GBM, GAM, CTA, ANN, SRE, FDA, MARS, RF, MAXENT.Phillips, MAXENT.Phillips.2

Options

a BIOMOD.models.options object returned by the BIOMOD_ModelingOptions function

VarImport

(optional, default NULL)
An integer corresponding to the number of permutations to be done for each variable to estimate variable importance

mod.eval.method

a vector containing evaluation metric names to be used, must be among ROC, TSS, KAPPA, ACCURACY, BIAS, POD, FAR, POFD, SR, CSI, ETS, HK, HSS, OR, ORSS

SavePred

(optional, default TRUE)
A logical value defining whether all results and outputs must be saved on hard drive or not (! strongly recommended !)

scal.models

(optional, default FALSE)
A logical value defining whether all models predictions must be scaled with a binomial GLM or not

Value

A list containing for each model a list containing the following elements :

ModelName : the name of correctly computed model

calib.failure : the name of incorrectly computed model

pred : the prediction outputs for calibration data

pred.eval : the prediction outputs for validation data

evaluation : the evaluation outputs returned by the bm_FindOptimStat function

var.import : the mean of variables importance returned by the bm_VariablesImportance function

See also

Author

Damien Georges