Sample binary vector

This internal biomod2 function allows the user to sample a binary vector keeping the same proportion of 0 and 1 as the initial vector.

bm_SampleBinaryVector(ref, ratio, as.logi = FALSE)

Arguments

ref
a vector containing binary values (either 0 or 1)

ratio
a numeric between 0 and 1 corresponding to the proportion of ref values to sample

as.logi
(optional, default FALSE)
A logical value defining whether a vector of TRUE/FALSE values should be returned (if as.logi = TRUE) or a vector containing the indices of ref elements to keep

Value

A list containing to elements is returned :

calibration

IDs of elements selected for calibration

evaluation

IDs of elements selected for evaluation (complementary to the calibration set)

Details

###############################################################################################

See also

Author

Damien Georges

Examples

## generate a binary vector
a <- sample(c(0,1), 100, replace = TRUE)

bm_SampleBinaryVector(ref = a, ratio = 0.7)

###############################################################################################