Variables' importance calculation

This internal biomod2 function allows the user to compute a variable importance value for each variable involved in the given model.

bm_VariablesImportance(
  bm.model,
  expl.var,
  method = "full_rand",
  nb.rep = 1,
  seed.val = NULL,
  ...
)

Arguments

bm.model

a biomod2_model object (or nnet, rpart, fda, gam, glm, lm, gbm, mars, randomForest) that can be obtained with the get_formal_model function

expl.var

a data.frame containing the explanatory variables that will be used to compute the variables importance

method

a character corresponding to the randomisation method to be used, must be full_rand (only method available so far)

nb.rep

an integer corresponding to the number of permutations to be done for each variable

seed.val

(optional, default NULL)
An integer value corresponding to the new seed value to be set

...

(optional))

Value

A matrix containing variables importance scores for each permutation run.

Details

For each variable to be evaluated :

shuffle the original variable

compute model prediction with shuffled variable

calculate Pearson's correlation between reference and shuffled predictions

return score as 1 - cor

The highest the value, the less reference and shuffled predictions are correlated, and the more influence the variable has on the model. A value of 0 assumes no influence of the variable on the model.

Note that this calculation does not account for variables' interactions.

The same principle is used in randomForest.

See also

Author

Damien Georges

Examples

## Create simple simulated data
myResp.s <- sample(c(0, 1), 20, replace = TRUE)
myExpl.s <- data.frame(var1 = sample(c(0, 1), 100, replace = TRUE),
                       var2 = rnorm(100),
                       var3 = 1:100)

## Compute variables importance
mod <- glm(var1 ~ var2 + var3, data = myExpl.s)
bm_VariablesImportance(bm.model = mod, 
                       expl.var = myExpl.s[, c('var2', 'var3')],
                       method = "full_rand",
                       nb.rep = 3)