Functions to extract informations from BIOMOD.models.out, BIOMOD.projection.out or BIOMOD.ensemble.models.out objects

These functions allow the user to easily retrieve informations stored in the different biomod2 objects from the different modeling steps, such as modeling options and formated data, models used or not, predictions, evaluations, variables importance.

# S4 method for BIOMOD.models.out
get_options(obj)

# S4 method for BIOMOD.models.out
get_calib_lines(obj, as.data.frame = FALSE, ...)

# S4 method for BIOMOD.models.out
get_formal_data(obj, subinfo = NULL)

# S4 method for BIOMOD.models.out
get_predictions(obj, as.data.frame = FALSE, evaluation = FALSE)

# S4 method for BIOMOD.models.out
get_built_models(obj, ...)

# S4 method for BIOMOD.models.out
get_evaluations(obj, as.data.frame = FALSE, ...)

# S4 method for BIOMOD.models.out
get_variables_importance(obj, as.data.frame = FALSE, ...)

# S4 method for BIOMOD.projection.out
get_projected_models(obj)

# S4 method for BIOMOD.projection.out
free(obj)

# S4 method for BIOMOD.projection.out
get_predictions(
  obj,
  as.data.frame = FALSE,
  full.name = NULL,
  model = NULL,
  run.eval = NULL,
  data.set = NULL
)

# S4 method for BIOMOD.ensemble.models.out
get_needed_models(obj, selected.models = "all", ...)

# S4 method for BIOMOD.ensemble.models.out
get_kept_models(obj, model, ...)

# S4 method for BIOMOD.ensemble.models.out
get_predictions(obj, ...)

# S4 method for BIOMOD.ensemble.models.out
get_built_models(obj, ...)

# S4 method for BIOMOD.ensemble.models.out
get_evaluations(obj, as.data.frame = FALSE, ...)

# S4 method for BIOMOD.ensemble.models.out
get_variables_importance(obj, as.data.frame = FALSE, ...)

Arguments

obj

a BIOMOD.models.out, BIOMOD.projection.out or BIOMOD.ensemble.models.out object

Author

Damien Georges