Functions to get predictions from biomod2_model objects

Internal S4 method used to predict single models from biomod2_model on (new) explanatory variables. predict2 was introduced to allow a signature with two arguments : object, a type of biomod2_model and newdata, either a SpatRaster or a data.frame.

predict2(object, newdata, ...)

# S4 method for biomod2_model,SpatRaster
predict2(object, newdata, predfun, seedval = NULL, ...)

# S4 method for biomod2_model,data.frame
predict2(object, newdata, predfun, seedval = NULL, ...)

# S4 method for ANN_biomod2_model,SpatRaster
predict2(object, newdata, ...)

# S4 method for ANN_biomod2_model,data.frame
predict2(object, newdata, ...)

# S4 method for CTA_biomod2_model,SpatRaster
predict2(object, newdata, ...)

# S4 method for CTA_biomod2_model,data.frame
predict2(object, newdata, ...)

# S4 method for FDA_biomod2_model,SpatRaster
predict2(object, newdata, ...)

# S4 method for FDA_biomod2_model,data.frame
predict2(object, newdata, ...)

# S4 method for GAM_biomod2_model,SpatRaster
predict2(object, newdata, ...)

# S4 method for GAM_biomod2_model,data.frame
predict2(object, newdata, ...)

# S4 method for GBM_biomod2_model,SpatRaster
predict2(object, newdata, ...)

# S4 method for GBM_biomod2_model,data.frame
predict2(object, newdata, ...)

# S4 method for GLM_biomod2_model,SpatRaster
predict2(object, newdata, ...)

# S4 method for GLM_biomod2_model,data.frame
predict2(object, newdata, ...)

# S4 method for MARS_biomod2_model,SpatRaster
predict2(object, newdata, ...)

# S4 method for MARS_biomod2_model,data.frame
predict2(object, newdata, ...)

# S4 method for MAXENT.Phillips_biomod2_model,SpatRaster
predict2(object, newdata, ...)

# S4 method for MAXENT.Phillips_biomod2_model,data.frame
predict2(object, newdata, do_raster = FALSE, newraster = NULL, ...)

# S4 method for MAXENT.Phillips.2_biomod2_model,SpatRaster
predict2(object, newdata, ...)

# S4 method for MAXENT.Phillips.2_biomod2_model,data.frame
predict2(object, newdata, ...)

# S4 method for RF_biomod2_model,SpatRaster
predict2(object, newdata, ...)

# S4 method for RF_biomod2_model,data.frame
predict2(object, newdata, ...)

# S4 method for SRE_biomod2_model,SpatRaster
predict2(object, newdata, ...)

# S4 method for SRE_biomod2_model,data.frame
predict2(object, newdata, ...)

Arguments

object

a biomod2_model object

newdata

a data.frame or SpatRaster object containing data for new predictions

predfun

a function, generated by the predict2 method specific to each biomod2_model subclass and used within the generic predict2.biomod2_model.SpatRaster or predict2.biomod2_model.data.frame to do the prediction

seedval

(optional, default NULL)
An integer value corresponding to the new seed value to be set

do_raster

(optional, default NULL). A boolean used with MAXENT.Phillips, so that the SpatRaster method can use the data.frame method.

newraster

(optional, default NULL). A [terra:SpatRaster]{SpatRaster} used by the MAXENT.Phillips method so that the data.frame method can transform its results back into SpatRaster when do_raster = TRUE.

...

(optional))

See also

Author

Remi Patin