Analysis of the response curves of a model within Biomod

Depreciated function, please use response.plot2 instead

response.plot(
  model,
  Data,
  show.variables = seq(1:ncol(Data)),
  save.file = "no",
  name = "response_curve",
  ImageSize = 480,
  plot = TRUE
)

Arguments

model

the model for which you want the response curves to be plotted. Compatible with GAM, GBM, GLM, ANN, CTA, RF, FDA and MARS.

Data

the variables for which you want the response curves to be plotted. A data frame is wanted with one column per variable. They have to have the same names as the ones used to calibrate the model.

show.variables

give in the column numbers of 'Data' for selecting the variables that are wanted for plotting

save.file

can be set to "pdf", "jpeg" or "tiff" to save the plot. Pdf options can be changed by setting the default values of pdf.options().

name

the name of the file produced if save.file is different to "no" (extensions are already included)

ImageSize

the size of the image in pixels if save.file is different to "no". Affects "jpeg" and "tiff" outputs only. Default if 480 pixels which is the R default.

plot

if TRUE (the default) then a plot is produced. If not, an array containing predictions is returned (see details)

Details

Depreciated function, please use response.plot2 instead.

References

Elith, J., Ferrier, S., Huettmann, FALSE. & Leathwick, J. R. 2005 The evaluation strip: A new and robust method for plotting predicted responses from species distribution models. Ecological Modelling 186, 280-289.

Author

Wilfried Thuiller