Run a range of species distribution models

This function allows to calibrate and evaluate a range of modeling techniques for a given species distribution. The dataset can be split up for independent calibration and validation, and the predictive power of the different models can be estimated using a range of evaluation metrics (see Details).

BIOMOD_Modeling(
  data,
  modeling.id = as.character(format(Sys.time(), "%s")),
  models = c("GLM", "GBM", "GAM", "CTA", "ANN", "SRE", "FDA", "MARS", "RF",
    "MAXENT.Phillips", "MAXENT.Phillips.2"),
  models.options = NULL,
  NbRunEval = 1,
  DataSplit = 100,
  Yweights = NULL,
  Prevalence = NULL,
  VarImport = 0,
  models.eval.meth = c("KAPPA", "TSS", "ROC"),
  SaveObj = TRUE,
  rescal.all.models = FALSE,
  do.full.models = TRUE,
  ...
)

Arguments

data

a BIOMOD.formated.data object returned by the BIOMOD_FormatingData function

modeling.id

a character corresponding to the name (ID) of the simulation set (a random number by default)

models

a vector containing model names to be computed, must be among GLM, GBM, GAM, CTA, ANN, SRE, FDA, MARS, RF, MAXENT.Phillips, MAXENT.Phillips.2

models.options

a BIOMOD.models.options object returned by the BIOMOD_ModelingOptions function

NbRunEval

an integer corresponding to the number of repetitions to be done for calibration/validation splitting
(if specified, DataSplit and do.full.models will be ignored)

DataSplit

a numeric between 0 and 1 corresponding to the percentage of data used to calibrate the models (calibration/validation splitting)

Yweights

(optional, default NULL)
A vector of numeric values corresponding to observation weights (one per observation, see Details)

Prevalence

(optional, default NULL)
A numeric between 0 and 1 corresponding to the species prevalence to build 'weighted response weights' (see Details)

VarImport

(optional, default NULL)
An integer corresponding to the number of permutations to be done for each variable to estimate variable importance

models.eval.meth

a vector containing evaluation metric names to be used, must be among ROC, TSS, KAPPA, ACCURACY, BIAS, POD, FAR, POFD, SR, CSI, ETS, HK, HSS, OR, ORSS

SaveObj

(optional, default TRUE)
A logical value defining whether all results and outputs must be saved on hard drive or not (! strongly recommended !)

rescal.all.models

(optional, default FALSE)
A logical value defining whether all models predictions must be scaled with a binomial GLM or not

do.full.models

(optional, default TRUE)
A logical value defining whether models calibrated and evaluated over the whole dataset must be computed or not

DataSplitTable

(optional, default NULL)
A matrix or data.frame defining for each repetition (in columns) which observation lines should be used for models calibration (TRUE) and validation (FALSE) (see BIOMOD_CrossValidation)

Value

A BIOMOD.models.out object containing models outputs, or links to saved outputs.
Models outputs are stored out of R (for memory storage reasons) in 2 different folders created in the current working directory :

a models folder, named after the resp.name argument of BIOMOD_FormatingData, and containing all calibrated models for each repetition and pseudo-absence run

a hidden folder, named .BIOMOD_DATA, and containing outputs related files (original dataset, calibration lines, pseudo-absences selected, predictions, variables importance, evaluation values...), that can be retrieved with get_[...] or load functions, and used by other biomod2 functions, like BIOMOD_Projection or BIOMOD_EnsembleModeling

Details

data

If you have decided to add pseudo absences to your original dataset (see BIOMOD_FormatingData),
PA.nb.rep *(NbRunEval + 1) models will be created.

models

The set of models to be calibrated on the data. 10 modeling techniques are currently available :

GLM : Generalized Linear Model (glm)

GAM : Generalized Additive Model (gam, gam or bam)
(see BIOMOD_ModelingOptions for details on algorithm selection)

GBM : Generalized Boosting Model, or usually called Boosted Regression Trees (gbm)

CTA : Classification Tree Analysis (rpart)

ANN : Artificial Neural Network (nnet)

SRE : Surface Range Envelop or usually called BIOCLIM

FDA : Flexible Discriminant Analysis (fda)

MARS : Multiple Adaptive Regression Splines (earth)

RF : Random Forest (randomForest)

MAXENT.Phillips : Maximum Entropy (https://biodiversityinformatics.amnh.org/open_source/maxent)

MAXENT.Phillips.2 : Maximum Entropy (maxnet)

NbRunEval & DataSplit

Most simple method in machine learning to calibrate and evaluate a model is to split the original dataset in two, one to calibrate the model and the other one to evaluate it. The DataSplit argument defines the percentage of data that will be randomly selected and used for the calibration part, the remaining data constituting the evaluation part. This process is repeated NbRunEval times, to be sure not to include bias both in the modeling and evaluation parts.

Other validation methods are also available to the user :

evaluation dataset can be directly given to the BIOMOD_FormatingData function

DataSplitTable argument can be used and obtained from the BIOMOD_CrossValidation function

Yweights & Prevalence

More or less weight can be given to some specific observations.

If Yweights = Prevalence = NULL, each observation (presence or absence) will have the same weight, no matter the total number of presences and absences.

If Prevalence = 0.5, presences and absences will be weighted equally (i.e. the weighted sum of presences equals the weighted sum of absences).

If Prevalence is set below (above) 0.5, more weight will be given to absences (presences).

If Yweights is defined, Prevalence argument will be ignored, and each observation will have its own weight.

If pseudo-absences have been generated (PA.nb.rep > 0 in BIOMOD_FormatingData), weights are by default calculated such that Prevalence = 0.5. Automatically created Yweights will be integer values to prevent some modeling issues.

models.eval.meth

ROC : Relative Operating Characteristic

KAPPA : Cohen's Kappa (Heidke skill score)

TSS : True kill statistic (Hanssen and Kuipers discriminant, Peirce's skill score)

FAR : False alarm ratio

SR : Success ratio

ACCURANCY : Accuracy (fraction correct)

BIAS : Bias score (frequency bias)

POD : Probability of detection (hit rate)

CSI : Critical success index (threat score)

ETS : Equitable threat score (Gilbert skill score)

Optimal value of each method can be obtained with the get_optim_value function. Several evaluation metrics can be selected. Please refer to the CAWRC website to get detailed description of each metric.

SaveObj

If this argument is set to FALSE, it may prevent the evaluation of the ensemble models (see BIOMOD_EnsembleModeling) in further steps. Strong recommandation is to keep SaveObj = TRUE, even if it requires to have some free space onto the hard drive.

rescal.all.models

This parameter is quite experimental and it is recommended not to use it. It may lead to reduction in projection scale amplitude. Some categorical models always have to be scaled (FDA, ANN), but it may be interesting to scale all computed models to ensure comparable predictions (0-1000 range). It might be particularly useful when doing ensemble forecasting to remove the scale prediction effect (the more extended projections are, the more they influence ensemble forecasting results).

do.full.models

Building models with all available information may be useful in some particular cases (e.g. rare species with few presences points). But calibration and evaluation datasets will be the same, which might lead to over-optimistic evaluation scores.

See also

Author

Wilfried Thuiller, Damien Georges, Robin Engler

Examples

# species occurrences
myFile <- system.file("external/species/mammals_table.csv", package="biomod2")
DataSpecies <- read.csv(myFile, row.names = 1)
head(DataSpecies)

# the name of studied species
myRespName <- 'GuloGulo'

# the presence/absences data for our species
myResp <- as.numeric(DataSpecies[, myRespName])

# the XY coordinates of species data
myRespXY <- DataSpecies[, c("X_WGS84", "Y_WGS84")]

# Environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
myFiles = paste0("external/bioclim/current/bio", c(3, 4, 7, 11, 12), ".grd")
myExpl = raster::stack(system.file(myFiles[1], package = "biomod2"),
                       system.file(myFiles[2], package = "biomod2"),
                       system.file(myFiles[3], package = "biomod2"),
                       system.file(myFiles[4], package = "biomod2"),
                       system.file(myFiles[5], package = "biomod2"))

# 1. Formating Data
myBiomodData <- BIOMOD_FormatingData(resp.var = myResp,
                                     expl.var = myExpl,
                                     resp.xy = myRespXY,
                                     resp.name = myRespName)

# 2. Defining Models Options using default options.
myBiomodOption <- BIOMOD_ModelingOptions()

# 3. Doing Modelisation
myBiomodModelOut <- BIOMOD_Modeling(myBiomodData,
                                    models = c('SRE','RF'),
                                    models.options = myBiomodOption,
                                    NbRunEval = 2,
                                    DataSplit = 80,
                                    VarImport = 3,
                                    models.eval.meth = c('TSS','ROC'),
                                    do.full.models = FALSE,
                                    modeling.id = 'test')

# print a summary of modeling stuff
myBiomodModelOut