Evaluate models with presence-only metrics

This function computes presence-only evaluation metrics (Boyce index and Minimal Predicted Area) for BIOMOD.models.out or BIOMOD.ensemble.models.out objects that can be obtained from BIOMOD_Modeling or BIOMOD_EnsembleModeling functions.

BIOMOD_PresenceOnly(
  modeling.output = NULL,
  EM.output = NULL,
  bg.env = NULL,
  perc = 0.9,
  save.output = TRUE
)

Arguments

modeling.output

(optional, default NULL)
A BIOMOD.models.out object that can be obtained from BIOMOD_Modeling function

EM.output

(optional, default NULL)
A BIOMOD.ensemble.models.out object that can be obtained from BIOMOD_EnsembleModeling function

bg.env

(optional, default NULL)
A matrix or data.frame object containing values of environmental variables extracted from the background (if presences are to be compared to background instead of absences or pseudo-absences selected for modeling)

perc

a numeric between 0 and 1 corresponding to the percentage of correctly classified presences for Minimal Predicted Area (see ecospat.mpa)

save.output

(optional, default TRUE)
A logical value defining whether the output is to be saved within the .BIOMOD_DATA folder or not

Value

A data.frame containing evaluation scores both for the evaluation metrics used in the BIOMOD_Modeling function and additional Boyce index and Minimal Predicted Area.

Details

em.by parameter of BIOMOD_EnsembleModeling must have been set to PA_dataset+repet in order to have an ensemble for each RUN of the NbRunEval parameter of the BIOMOD_Modeling function for evaluation.

The Boyce index returns NA values for SRE models because it can not be calculated with binary predictions.
This is also the reason why some NA values might appear for GLM models if they do not converge.

References

Engler, R., Guisan, A., and Rechsteiner L. 2004. An improved approach for predicting the distribution of rare and endangered species from occurrence and pseudo-absence data. Journal of Applied Ecology, 41(2), 263-274.

Hirzel, A. H., Le Lay, G., Helfer, V., Randin, C., and Guisan, A. 2006. Evaluating the ability of habitat suitability models to predict species presences. Ecological Modelling, 199(2), 142-152.

See also

Author

Frank Breiner, Maya Gueguen

Examples

# species occurrences
myFile <- system.file("external/species/mammals_table.csv", package="biomod2")
DataSpecies <- read.csv(myFile, row.names = 1)
head(DataSpecies)

# the name of studied species
myRespName <- 'GuloGulo'

# the presence/absences data for our species
myResp <- as.numeric(DataSpecies[, myRespName])

# the XY coordinates of species data
myRespXY <- DataSpecies[, c("X_WGS84", "Y_WGS84")]

# Environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
myFiles = paste0("external/bioclim/current/bio", c(3, 4, 7, 11, 12), ".grd")
myExpl = raster::stack(system.file(myFiles[1], package = "biomod2"),
                       system.file(myFiles[2], package = "biomod2"),
                       system.file(myFiles[3], package = "biomod2"),
                       system.file(myFiles[4], package = "biomod2"),
                       system.file(myFiles[5], package = "biomod2"))

# 1. Formatting Data
myBiomodData <- BIOMOD_FormatingData(resp.var = myResp,
                                     expl.var = myExpl,
                                     resp.xy = myRespXY,
                                     resp.name = myRespName)

# 2. Defining Models Options using default options.
myBiomodOption <- BIOMOD_ModelingOptions()

# 3. Doing Modelisation
myBiomodModelOut <- BIOMOD_Modeling(myBiomodData, 
                                    models = c('SRE', 'CTA', 'RF'), 
                                    models.options = myBiomodOption, 
                                    NbRunEval = 1, 
                                    DataSplit = 80, 
                                    Yweights = NULL, 
                                    VarImport = 3, 
                                    models.eval.meth = c('TSS','ROC'),
                                    SaveObj = TRUE,
                                    rescal.all.models = FALSE,
                                    do.full.models = FALSE)

# 4. Doing Ensemble Modelling
myBiomodEM <- BIOMOD_EnsembleModeling(modeling.output = myBiomodModelOut,
                                      chosen.models = 'all',
                                      em.by = 'PA_dataset+repet',
                                      eval.metric = c('TSS'),
                                      eval.metric.quality.threshold = c(0.7),
                                      models.eval.meth = c('TSS','ROC'),
                                      prob.mean = TRUE,
                                      prob.cv = FALSE,
                                      prob.ci = FALSE,
                                      prob.ci.alpha = 0.05,
                                      prob.median = FALSE,
                                      committee.averaging = FALSE,
                                      prob.mean.weight = TRUE,
                                      prob.mean.weight.decay = 'proportional')  

requireNamespace(PresenceAbsence, 'PresenceAbsence', quietly = TRUE)

# evaluate Biomod models with the Boyce index and MPA
pres.only.eval <- BIOMOD_PresenceOnly(myBiomodModelOut, myBiomodEM)
pres.only.eval$eval

# evaluate Biomod models with the Boyce index and MPA using Background data
bg.Values <- getValues(myExpl)
pres.only.eval <- BIOMOD_PresenceOnly(myBiomodModelOut, myBiomodEM, bg.env = bg.Values)
pres.only.eval$eval