Analyze the range size differences between projections of species distribution models

This function allows to calculate the absolute number of locations (pixels) lost, stable and gained, as well as the corresponding relative proportions, between two (or more) binary projections of (ensemble) species distribution models (which can represent new time scales or environmental scenarios for example).

BIOMOD_RangeSize(proj.current, proj.future)

# S4 method for data.frame,data.frame
BIOMOD_RangeSize(proj.current, proj.future)

# S4 method for SpatRaster,SpatRaster
BIOMOD_RangeSize(proj.current, proj.future)

Arguments

proj.current

a data.frame, RasterLayer or SpatRaster object containing the initial binary projection(s) of the (ensemble) species distribution model(s)

proj.future

a data.frame, RasterLayer or SpatRaster object containing the final binary projection(s) of the (ensemble) species distribution model(s)

Value

A list containing two objects :

Compt.By.Species

a data.frame containing the summary of range change for each comparison

Loss : number of pixels predicted to be lost

Stable0 : number of pixels not currently occupied and not predicted to be

Stable1 : number of pixels currently occupied and predicted to remain occupied

Gain : number of pixels predicted to be gained

PercLoss : percentage of pixels currently occupied and predicted to be lost (Loss / (Loss + Stable1))

PercGain : percentage of pixels predicted to be gained compare to the number of pixels currently occupied (Gain / (Loss + Stable1))

SpeciesRangeChange : percentage of pixels predicted to change (loss or gain) compare to the number of pixels currently occupied (PercGain - PercLoss)

CurrentRangeSize : number of pixels currently occupied

FutureRangeSize0Disp : number of pixels predicted to be occupied, assuming no migration

FutureRangeSize1Disp : number of pixels predicted to be occupied, assuming migration

Diff.By.Pixel

an object in the same form than the input data (proj.current and proj.future) and containing a value for each point/pixel of each comparison among :

-2 : predicted to be lost

-1 : predicted to remain occupied

0 : predicted to remain unoccupied

1 : predicted to be gained

Details

Note that this function is only relevant to compare binary projections, made on the same area with the same resolution.

Comparison between proj.current and proj.future depends on the number of projection in both objects:

proj.current	proj.future	Comparison

1 projection (e.g. data.frame with 1 column, SpatRaster with 1 layer)	1 projection (e.g. data.frame with 1 column, SpatRaster with 1 layer)	comparison of both projection (e.g. current vs future conditions for the same model ; current vs current condition for two different models)

n projections (e.g. data.frame with n column, SpatRaster with n layer)	n projections (e.g. data.frame with n column, SpatRaster with n layer)	comparing projection i in proj.current to projection i in proj.future (e.g. comparing current vs future condition for n models)

1 projection (e.g. data.frame with 1 column, SpatRaster with 1 layer)	n projections (e.g. data.frame with n column, SpatRaster with n layer)	comparing projection in proj.current to each projection in proj.future (e.g. comparing current vs n different future condition (e.g. climate change scenario) for 1 model)

Diff.By.Pixel object is obtained by applying the simple following formula : $$proj.future - 2 * proj.current$$

See also

Author

Wilfried Thuiller, Damien Georges, Bruno Lafourcade

Examples

library(terra)

# Load species occurrences (6 species available)
data(DataSpecies)
head(DataSpecies)

# Select the name of the studied species
myRespName <- 'GuloGulo'

# Get corresponding presence/absence data
myResp <- as.numeric(DataSpecies[, myRespName])

# Get corresponding XY coordinates
myRespXY <- DataSpecies[, c('X_WGS84', 'Y_WGS84')]

# Load environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
data(bioclim_current)
myExpl <- terra::rast(bioclim_current)

# \dontshow{
myExtent <- terra::ext(0,30,45,70)
myExpl <- terra::crop(myExpl, myExtent)
# }

# --------------------------------------------------------------- #
file.out <- paste0(myRespName, "/", myRespName, ".AllModels.models.out")
if (file.exists(file.out)) {
  myBiomodModelOut <- get(load(file.out))
} else {

  # Format Data with true absences
  myBiomodData <- BIOMOD_FormatingData(resp.var = myResp,
                                       expl.var = myExpl,
                                       resp.xy = myRespXY,
                                       resp.name = myRespName)

  # Create default modeling options
  myBiomodOptions <- BIOMOD_ModelingOptions()

  # Model single models
  myBiomodModelOut <- BIOMOD_Modeling(bm.format = myBiomodData,
                                      modeling.id = 'AllModels',
                                      models = c('RF', 'GLM'),
                                      bm.options = myBiomodOptions,
                                      nb.rep = 2,
                                      data.split.perc = 80,
                                      metric.eval = c('TSS','ROC'),
                                      var.import = 3,
                                      do.full.models = FALSE,
                                      seed.val = 42)
}

file.proj <- paste0(myRespName, "/proj_Current/", myRespName, ".Current.projection.out")
if (file.exists(file.proj)) {
  myBiomodProj <- get(load(file.proj))
} else {

  # Project single models
  myBiomodProj <- BIOMOD_Projection(bm.mod = myBiomodModelOut,
                                    proj.name = 'Current',
                                    new.env = myExpl,
                                    models.chosen = 'all',
                                    metric.binary = 'all',
                                    metric.filter = 'all',
                                    build.clamping.mask = TRUE)
}

# --------------------------------------------------------------- #
# Load environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
data(bioclim_future)
myExplFuture <- terra::rast(bioclim_future)
# \dontshow{
myExtent <- terra::ext(0,30,45,70)
myExplFuture <- terra::crop(myExplFuture, myExtent)
# }
# Project onto future conditions
myBiomodProjectionFuture <- BIOMOD_Projection(bm.mod = myBiomodModelOut,
                                              proj.name = 'Future',
                                              new.env = myExplFuture,
                                              models.chosen = 'all',
                                              metric.binary = 'TSS',
                                              build.clamping.mask = TRUE)

# Load current and future binary projections
proj.current <- terra::rast("GuloGulo/proj_Current/proj_Current_GuloGulo_TSSbin.tif")
proj.future <- terra::rast("GuloGulo/proj_Future/proj_Future_GuloGulo_TSSbin.tif")

# Compute differences
myBiomodRangeSize <- BIOMOD_RangeSize(proj.current = proj.current, proj.future = proj.future)

myBiomodRangeSize$Compt.By.Models
plot(myBiomodRangeSize$Diff.By.Pixel)

# Represent main results 
bm_PlotRangeSize(bm.range = myBiomodRangeSize)