Tune models parameters

Function to tune biomod2 single models parameters

BIOMOD_Tuning(
  data,
  models = c("GLM", "GBM", "GAM", "CTA", "ANN", "SRE", "FDA", "MARS", "RF",
    "MAXENT.Phillips"),
  models.options = BIOMOD_ModelingOptions(),
  method.ANN = "avNNet",
  method.RF = "rf",
  method.MARS = "earth",
  method.GAM = "gam",
  method.GLM = "glmStepAIC",
  trControl = NULL,
  metric = "ROC",
  ctrl.CTA = NULL,
  ctrl.RF = NULL,
  ctrl.ANN = NULL,
  ctrl.MARS = NULL,
  ctrl.FDA = NULL,
  ctrl.GAM = NULL,
  ctrl.GBM = NULL,
  ctrl.GLM = NULL,
  tuneLength = 30,
  decay.tune.ANN = c(0.001, 0.01, 0.05, 0.1),
  size.tune.ANN = c(2, 4, 6, 8),
  maxit.ANN = 500,
  MaxNWts.ANN = 10 * (ncol(data@data.env.var) + 1) + 10 + 1,
  type.GLM = c("simple", "quadratic", "polynomial", "s_smoother"),
  interaction.GLM = c(0, 1),
  cvmethod.ME = "randomkfold",
  overlap.ME = FALSE,
  kfolds.ME = 10,
  n.bg.ME = 10000,
  env.ME = NULL,
  metric.ME = "ROC",
  clamp.ME = TRUE,
  parallel.ME = FALSE,
  numCores.ME = NULL,
  Yweights = NULL
)

Arguments

data

a BIOMOD.formated.data object returned by the BIOMOD_FormatingData function

models

a vector containing model names to be tuned,
must be among GLM, GBM, GAM, CTA, ANN, SRE, FDA, MARS, RF, MAXENT.Phillips

models.options

a BIOMOD.models.options object returned by the BIOMOD_ModelingOptions function

method.ANN

a character corresponding to the classification or regression model to use for ANN,
must be avNNet (see http://topepo.github.io/caret/Neural_Network.html)

method.RF

a character corresponding to the classification or regression model to use for RF,
must be rf (see http://topepo.github.io/caret/Random_Forest.html)

method.MARS

a character corresponding to the classification or regression model to use for MARS,
must be earth (see http://topepo.github.io/caret/Multivariate_Adaptive_Regression_Splines.html)

method.GAM

a character corresponding to the classification or regression model to use for GAM,
must be gam (see http://topepo.github.io/caret/Generalized_Additive_Model.html)

method.GLM

a character corresponding to the classification or regression model to use for GLM,
must be glmStepAIC (see http://topepo.github.io/caret/Generalized_Linear_Model.html)

trControl

global control parameters that can be obtained from trainControl function

metric

a character corresponding to the evaluation metric used to select optimal models and tune parameters, must be either ROC or TSS (maximizing Sensitivity and Specificity)

ctrl.CTA

control parameters for CTA

ctrl.RF

control parameters for RF

ctrl.ANN

control parameters for ANN

ctrl.MARS

control parameters for MARS

ctrl.FDA

control parameters for FDA

ctrl.GAM

control parameters for GAM

ctrl.GBM

control parameters for GBM

ctrl.GLM

control parameters for GLM

tuneLength

(see train)

decay.tune.ANN

a vector of weight decay parameters for ANN

size.tune.ANN

a vector of size parameters (number of units in the hidden layer) for ANN

maxit.ANN

an integer corresponding to the maximum number of iterations for ANN

MaxNWts.ANN

an integer corresponding to the maximum allowable number of weights for ANN

type.GLM

a vector of character corresponding to modeling types for GLM,
must be among simple, quadratic, polynomial, s_smoother

interaction.GLM

a vector of interaction types, must be amond 0, 1

cvmethod.ME

a character corresponding to the method used to partition data for MAXENT.Phillips,
must be randomkfold

overlap.ME

(optional, default FALSE)
A logical value defining whether to calculate pairwise metric of niche overlap or not (see calc.niche.overlap)

kfolds.ME

an integer corresponding to the number of bins for k-fold cross-validation for MAXENT.Phillips

n.bg.ME

an integer corresponding to the number of background points used to run MAXENT.Phillips

env.ME

a RasterStack object containing model predictor variables

metric.ME

a character corresponding to the evaluation metric used to select optimal model and tune parameters for MAXENT.Phillips, must be either auc.val.avg, auc.diff.avg, or.mtp.avg, or.10p.avg or AICc

clamp.ME

(optional, default TRUE)
A logical value defining whether Features are constrained to remain within the range of values in the training data (Elith et al. 2011) or not

parallel.ME

(optional, default TRUE)
A logical value defining whether to enable parallel computing for MAXENT.Phillips or not

numCores.ME

an integer corresponding to the number of cores to be used to train MAXENT.Phillips

Yweights

a vector of response points weights for models allowing case weights

Value

A BIOMOD.models.options object (see BIOMOD_ModelingOptions) with optimized parameters

Details

trControl parameter is set by default to :
caret::trainControl(method = 'cv', summaryFunction = caret::twoClassSummary,
classProbs = TRUE, returnData = FALSE).

All control parameters for other models are set to trControl if unspecified.

For more details on MAXENT.Phillips tuning, please refer to ENMevaluate.

For more details on other models tuning, please refer to train.

References

Kuhn, Max. 2008. Building predictive models in R using the caret package. Journal of Statistical Software 28, 1-26.

Kuhn, Max, and Kjell Johnson. 2013. Applied predictive modeling. New York: Springer.

Muscarella, Robert, et al. 2014. ENMeval: An R package for conducting spatially independent evaluations and estimating optimal model complexity for Maxent ecological niche models. Methods in Ecology and Evolution, 5, 1198-1205.

See also

Author

Frank Breiner

Examples

# species occurrences
myFile <- system.file("external/species/mammals_table.csv", package="biomod2")
DataSpecies <- read.csv(myFile, row.names = 1)
head(DataSpecies)

# the name of studied species
myRespName <- 'GuloGulo'

# the presence/absences data for our species
myResp <- as.numeric(DataSpecies[, myRespName])

# the XY coordinates of species data
myRespXY <- DataSpecies[, c("X_WGS84", "Y_WGS84")]

# Environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
myFiles = paste0("external/bioclim/current/bio", c(3, 4, 7, 11, 12), ".grd")
myExpl = raster::stack(system.file(myFiles[1], package = "biomod2"),
                       system.file(myFiles[2], package = "biomod2"),
                       system.file(myFiles[3], package = "biomod2"),
                       system.file(myFiles[4], package = "biomod2"),
                       system.file(myFiles[5], package = "biomod2"))

# 1. Formating Data
myBiomodData <- BIOMOD_FormatingData(resp.var = myResp,
                                     expl.var = myExpl,
                                     resp.xy = myRespXY,
                                     resp.name = myRespName)

# 2. Defining Models Options using default options
### Duration for turing all models sequential with default settings 
### on 3.4 GHz processor: approx. 45 min tuning all models in parallel
### (on 8 cores) using foreach loops runs much faster: approx. 14 min

# library(doParallel)
# cl<-makeCluster(8)
# doParallel::registerDoParallel(cl) 

time.seq <- system.time(
  bm.tuning <- BIOMOD_Tuning(myBiomodData,
                             env.ME = myExpl,
                             n.bg.ME = ncell(myExpl))
)

# stopCluster(cl)

par(mfrow = c(1,3))
plot(bm.tuning$tune.CTA.rpart)
plot(bm.tuning$tune.CTA.rpart2)
plot(bm.tuning$tune.RF)

myBiomodModelOut <- BIOMOD_Modeling(myBiomodData, 
                                    models = c('RF', 'CTA'), 
                                    models.options = bm.tuning$models.options, 
                                    NbRunEval = 1, 
                                    DataSplit = 100, 
                                    VarImport = 0,
                                    models.eval.meth = c('ROC'),
                                    do.full.models = FALSE,
                                    modeling.id = 'test')