Convert species' probability of occurrence into binary presence-absence data using a predefined threshold

Function that converts an object containing probability values into a binary presence-absence object according to a pre-defined threshold(s).

.convertBin.matrix(data, threshold, doFiltering = FALSE)

Arguments

data
numeric vector, a matrix, a data.frame, a RasterLayer or a RasterStack containing the data to be converted

threshold
numeric value or a vector containing the threshold to be used for converting data.

Value

An object of the same class than data with binary (0 or 1) values, usually presence-absence.

Details

If data is a vector or a raster object, then the threshold should be a numeric value. If data is matrix,dataframe or rasterStack, then the threshold should have, in theory, as many values as the number of columns or layers to transform. In the particular case that the data to convert is a matrix/data.frame with several columns or a RasterStack with several layers and the threshold is a single numeric value, the same threshold will be applied to all columns (resp. layers).

Author

Wilfried Thuiller, Damien Georges

Examples

  xx <- rnorm(50,10)
  yy <- BinaryTransformation(xx, 10)

  cbind(xx,yy)