Standardized formula maker

This internal biomod2 function allows the user to create easily a standardized formula that can be used later by statistical models.

bm_MakeFormula(respName, explVar, type = "simple", interaction.level = 0, ...)

Arguments

respName

a character defining the response variable name

explVar

a matrix or data.frame corresponding to the explanatory variables table that will be considered at modelling step

type

a character defining the wanted type of formula, must be simple, quadratic, polynomial or s_smoother

interaction.level

an integer corresponding to the interaction level depth between explanatory variables

...

some additional arguments (see Details)

Value

A formula class object that can be directly given to most of Rstatistical models.

Details

It is advised to give only a subset of explVar table to avoid useless memory consuming.
If some explanatory variables are factorial, explVar must be a data.frame whose corresponding columns are defined as factor.

... can take the following values :

k : an integer corresponding to the smoothing parameter value of s or s arguments (used only if type = 's_smoother')

See also

Author

Damien Georges

Examples

## Create simulated data
myResp <- sample(c(0, 1), 20, replace = TRUE)
myExpl <- matrix(runif(60), 
                 ncol = 3,
                 dimnames = list(NULL, c('var1', 'var2', 'var3')))

## Create a formula
myFormula <- bm_MakeFormula(respName = 'myResp',
                            explVar = head(myExpl),
                            type = 'quadratic',
                            interaction.level = 0)
  
## Show formula created
myFormula