Sample binary vector

This internal biomod2 function allows the user to sample a binary vector keeping the same proportion of 0 and 1 as the initial vector.

bm_SampleBinaryVector(obs, ratio, as.logical = FALSE, seedval = NULL)

Arguments

obs

a vector containing binary values (either 0 or 1)

ratio

a numeric between 0 and 1 corresponding to the proportion of obs values to sample

as.logical

(optional, default FALSE)
A logical value defining whether output should be returned as a vector of TRUE/FALSE values or integer values corresponding to the indices of obs elements to be kept

seedval

(optional, default NULL)
An integer value corresponding to the new seed value to be set

Value

A list containing the following elements :

calibration : elements selected for calibration

validation : elements selected for validation (complementary to the calibration set)

See also

Author

Damien Georges

Examples

## Generate a binary vector
vec.a <- sample(c(0, 1), 100, replace = TRUE)

## Generate calibration / validation datasets
bm_SampleBinaryVector(obs = vec.a, ratio = 0.7)