Tool to ensure the sampling of all levels of a factorial variable

This internal biomod2 function samples randomly an element of each level of all the factorial variables contained in a raster* or data.frame object.

bm_SampleFactorLevels(x, mask.out = NULL, mask.in = NULL)

Arguments

x

a data.frame or RasterStack object containing the explanatory variables (in columns or layers)

mask.out

a data.frame or raster object containing the area that has already been sampled (factor levels within this mask will not be sampled)

mask.in

a data.frame, raster or RasterStack object containing areas where factor levels are to be sampled in priority. Note that if after having explored these masks, some factor levels remain unsampled, they will be sampled in the reference input object x.

Value

A numeric vector containing point IDs (either cell number for raster* objects, or row number for data.frame), each refering to a single level of a single factorial variable. In case any factorial variable is found in the input object, NULL is returned.

Details

The x, mask.out and mask.in parameters must be coherent in terms of dimensions :

same number of rows for data.frame objects

same resolution, projection system and number of cells for raster* objects

If mask.in contains several masks (either it is a RasterStack object or a multi-columns data.frame), then the order of masks / columns matters : they will be considered successively to sample missing factor levels.

raster* masks will be understood as :

NA : out of mask

not NA : in mask

data.frame masks will be understood as :

FALSE : out of mask

TRUE : in mask

Author

Damien Georges

Examples

## Example with raster* object ----------------------------------

library(raster)

## create a factorial raster
r1 <- raster()
r1[] <- 1; r1[1] <- 2; r1[2:3] <- 3
r1 <- as.factor(r1)

## create a continuous raster
r2 <- raster()
r2[] <- rnorm(ncell(r2))

## put the raster into a RasterStack
stk <- stack(r1, r2)
is.factor(stk)

## define a mask for already sampled points
mask.out <- r1
mask.out[] <- NA; mask.out[2:3] <- 1

## define a list of masks where we want to sample in priority
mask.in.1 <- mask.in.2 <- r1
mask.in.1[1:10] <- NA ## only level 1 should be sampled in this mask
mask.in.2[1] <- NA ## only levels 1 and 3 should be sampled in this mask
mask.in <- list(mask.in.1 = mask.in.1, 
                mask.in.2 = mask.in.2)

## test different version of the function
bm_SampleFactorLevels(stk, mask.out = mask.out)
bm_SampleFactorLevels(stk, mask.in = mask.in)
bm_SampleFactorLevels(stk, mask.out = mask.out, mask.in = mask.in)