Functions to extract informations from BIOMOD.models.out, BIOMOD.projection.out or BIOMOD.ensemble.models.out objects

These functions allow the user to easily retrieve informations stored in the different biomod2 objects from the different modeling steps, such as modeling options and formated data, models used or not, predictions, evaluations, variables importance.

# S4 method for BIOMOD.formated.data
get_species_data(obj)

# S4 method for BIOMOD.formated.data.PA
get_species_data(obj)

# S4 method for BIOMOD.formated.data
get_eval_data(obj)

# S4 method for BIOMOD.models.out
get_options(obj)

# S4 method for BIOMOD.models.out
get_calib_lines(obj, as.data.frame = FALSE, PA = NULL, run = NULL)

# S4 method for BIOMOD.models.out
get_formal_data(obj, subinfo = NULL)

# S4 method for BIOMOD.models.out
get_predictions(
  obj,
  evaluation = FALSE,
  full.name = NULL,
  PA = NULL,
  run = NULL,
  algo = NULL,
  model.as.col = FALSE
)

# S4 method for BIOMOD.models.out
get_built_models(obj, full.name = NULL, PA = NULL, run = NULL, algo = NULL)

# S4 method for BIOMOD.models.out
get_evaluations(
  obj,
  full.name = NULL,
  PA = NULL,
  run = NULL,
  algo = NULL,
  metric.eval = NULL
)

# S4 method for BIOMOD.models.out
get_variables_importance(
  obj,
  full.name = NULL,
  PA = NULL,
  run = NULL,
  algo = NULL,
  expl.var = NULL
)

# S4 method for BIOMOD.projection.out
get_projected_models(
  obj,
  full.name = NULL,
  PA = NULL,
  run = NULL,
  algo = NULL,
  merged.by.algo = NULL,
  merged.by.run = NULL,
  merged.by.PA = NULL,
  filtered.by = NULL
)

# S4 method for BIOMOD.projection.out
free(obj)

# S4 method for BIOMOD.projection.out
get_predictions(
  obj,
  metric.binary = NULL,
  metric.filter = NULL,
  full.name = NULL,
  PA = NULL,
  run = NULL,
  algo = NULL,
  merged.by.algo = NULL,
  merged.by.run = NULL,
  merged.by.PA = NULL,
  filtered.by = NULL,
  model.as.col = FALSE,
  ...
)

# S4 method for BIOMOD.ensemble.models.out
get_formal_data(obj, subinfo = NULL)

# S4 method for BIOMOD.ensemble.models.out
get_built_models(
  obj,
  full.name = NULL,
  merged.by.algo = NULL,
  merged.by.run = NULL,
  merged.by.PA = NULL,
  filtered.by = NULL,
  algo = NULL
)

# S4 method for BIOMOD.ensemble.models.out
get_kept_models(obj)

# S4 method for BIOMOD.ensemble.models.out
get_predictions(
  obj,
  evaluation = FALSE,
  full.name = NULL,
  merged.by.algo = NULL,
  merged.by.run = NULL,
  merged.by.PA = NULL,
  filtered.by = NULL,
  algo = NULL,
  model.as.col = FALSE
)

# S4 method for BIOMOD.ensemble.models.out
get_evaluations(
  obj,
  full.name = NULL,
  merged.by.algo = NULL,
  merged.by.run = NULL,
  merged.by.PA = NULL,
  filtered.by = NULL,
  algo = NULL,
  metric.eval = NULL
)

# S4 method for BIOMOD.ensemble.models.out
get_variables_importance(
  obj,
  full.name = NULL,
  merged.by.algo = NULL,
  merged.by.run = NULL,
  merged.by.PA = NULL,
  filtered.by = NULL,
  algo = NULL,
  expl.var = NULL
)

Arguments

obj

a BIOMOD.formated.data, BIOMOD.formated.data.PA, BIOMOD.models.out, BIOMOD.projection.out or BIOMOD.ensemble.models.out object

as.data.frame

a logical defining whether output should be returned as data.frame or array object

PA

(optional, default NULL)
A vector containing pseudo-absence set to be loaded, must be among PA1, PA2, ..., allData

run

(optional, default NULL)
A vector containing repetition set to be loaded, must be among RUN1, RUN2, ..., allRun

subinfo

a character corresponding to the information to be extracted, must be among NULL, expl.var.names, resp.var, expl.var, MinMax, eval.resp.var, eval.expl.var (see Details)

evaluation

a logical defining whether evaluation data should be used or not

full.name

(optional, default NULL)
A vector containing model names to be kept, must be either all or a sub-selection of model names that can be obtained with the get_built_models function

algo

(optional, default NULL)
A character containing algorithm to be loaded, must be either GLM, GBM, GAM, CTA, ANN, SRE, FDA, MARS, RF, MAXENT, MAXNET

model.as.col

(optional, default FALSE)
A boolean given to get_predictions. If TRUE prediction are returned as a wide data.frame with each column containing predictions for a single model and corresponding to the old output given by biomod2 in version < 4.2-2. If FALSE predictions are returned as a long data.frame with many additional informations readily available.

metric.eval

(optional, default NULL)
A vector containing evaluation metric to be kept, must be among ROC, TSS, KAPPA, ACCURACY, BIAS, POD, FAR, POFD, SR, CSI, ETS, HK, HSS, OR, ORSS

expl.var

(optional, default NULL)
A vector containing explanatory variables to be kept, that can be obtained with the get_formal_data(obj, subinfo = 'expl.var.names') function

merged.by.algo

(optional, default NULL)
A character containing merged algorithm to be loaded, must be among GLM, GBM, GAM, CTA, ANN, SRE, FDA, MARS, RF, MAXENT, MAXNET, mergedAlgo

merged.by.run

(optional, default NULL)
A vector containing merged repetition set to be loaded, must be among RUN1, RUN2, ..., mergedRun

merged.by.PA

(optional, default NULL)
A vector containing merged pseudo-absence set to be loaded, must be among PA1, PA2, ..., mergedData

filtered.by

(optional, default NULL)
A vector containing evaluation metric selected to filter single models to build the ensemble models, must be among ROC, TSS, KAPPA, ACCURACY, BIAS, POD, FAR, POFD, SR, CSI, ETS, HK, HSS, OR, ORSS

metric.binary

(optional, default NULL)
A vector containing evaluation metric selected to transform predictions into binary values, must be among ROC, TSS, KAPPA, ACCURACY, BIAS, POD, FAR, POFD, SR, CSI, ETS, HK, HSS, OR, ORSS

metric.filter

(optional, default NULL)
A vector containing evaluation metric to filter predictions, must be among ROC, TSS, KAPPA, ACCURACY, BIAS, POD, FAR, POFD, SR, CSI, ETS, HK, HSS, OR, ORSS

...

(optional, one or several of the following arguments depending on the selected function))

Value

get_species_data

a data.frame combining data.species, coord, data.env.var (and PA.table) slots of BIOMOD.formated.data (or BIOMOD.formated.data.PA) object

get_eval_data

a data.frame combining eval.data.species, eval.coord, eval.data.env.var slots of BIOMOD.formated.data or BIOMOD.formated.data.PA object

get_options

a BIOMOD.stored.models.options-class object from the models.options slot of a BIOMOD.models.out-class object

get_calib_lines

a BIOMOD.stored.data.frame-class object from the calib.lines slot of a BIOMOD.models.out object

get_projected_models

a vector from the models.projected slot of a BIOMOD.projection.out object

get_predictions

a BIOMOD.stored.data object from the proj.out slot of a BIOMOD.models.out, BIOMOD.projection.out or BIOMOD.ensemble.models.out object

get_kept_models

a vector containing names of the kept models of a BIOMOD.ensemble.models.out object

get_formal_data

depending on the subinfo parameter :

NULL

a BIOMOD.stored.formated.data-class (or BIOMOD.stored.models.out-class) object from the formated.input.data (or models.out) slot of a BIOMOD.models.out (or BIOMOD.ensemble.models.out) object

expl.var.names

a vector from the expl.var.names slot of a BIOMOD.models.out or BIOMOD.ensemble.models.out object

resp.var

a vector from the data.species slot of the formated.input.data slot of a BIOMOD.models.out or BIOMOD.ensemble.models.out object

expl.var

a data.frame from the data.env.var slot of the formated.input.data slot of a BIOMOD.models.out or BIOMOD.ensemble.models.out object

MinMax

a list of minimum and maximum values (or levels if factorial) of variable contained in the data.env.var slot of the formated.input.data slot of a BIOMOD.models.out or BIOMOD.ensemble.models.out object

eval.resp.var

a vector from the eval.data.species slot of the formated.input.data slot of a BIOMOD.models.out or BIOMOD.ensemble.models.out object

eval.expl.var

a data.frame from the eval.data.env.var slot of the formated.input.data slot of a BIOMOD.models.out or BIOMOD.ensemble.models.out object

get_built_models

a vector from the models.computed slot (or em.computed) of a BIOMOD.models.out (or BIOMOD.ensemble.models.out) object

get_evaluations

a data.frame from the models.evaluation slot (or model_evaluation of each model in em.computed) of a BIOMOD.models.out (or BIOMOD.ensemble.models.out) object. Contains evaluation metric for different models and dataset. Evaluation metric are calculated on the calibrating data (column calibration), on the cross-validation data (column validation) or on the evaluation data (column evaluation)

get_variables_importance

a BIOMOD.stored.data.frame-class from the variables.importance slot (or model_variables_importance of each model in em.models) of a BIOMOD.models.out (or BIOMOD.ensemble.models.out) object

See also

Author

Damien Georges